+++ bidict/__init__.py
@@ -9,7 +9,7 @@ and related functionality.
 
 """
 
-from ._common import BidirectionalMapping, BidictException, ValueExistsException
+from ._common import BidirectionalMapping, BidictException, KeyExistsException, ValueExistsException
 from ._bidict import bidict
 from ._loose import loosebidict
 from ._frozen import frozenbidict
@@ -19,6 +19,7 @@ from .util import pairs, inverted
 __all__ = (
     'BidirectionalMapping',
     'BidictException',
+    'KeyExistsException',
     'ValueExistsException',
     'bidict',
     'loosebidict',